pyproject.toml
[tool.poetry]
name = "speech-to-text"
version = "0.1.0"
description = ""
authors = ["eugepineiro <epineiro@itba.edu.ar>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.7.0"
transformers = "^4.36.2"
dataset = "^1.6.2"
datasets = "^2.16.1"
torch = "^2.1.2"
torchvision = "^0.16.2"
torchaudio = "^2.1.2"
soundfile = "^0.12.1"
librosa = "^0.10.1"
openai-whisper = "^20231117"
pydub = "^0.25.1"
jiwer = "^3.0.3"
accelerate = "^0.26.1"
nltk = "^3.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
